Fixed loading relative images
Saw a SO question linking to this repo. I wanted to know if I could use relative paths in my readme. This repo was marked as the accepted answer on SO but then stopped working. See http://stackoverflow.com/questions/10045517/embedding-images-inside-a-github-wiki-gollum-repository

I don't know how to add frames in markdown or even do alignment. So I've deleted that code. Just wanted to update the relative path example, and notify you that these examples no longer work.

Thanks